backend/requirements.txt
fastapi
uvicorn
tortoise-orm
aiosqlite
python-dotenv